Switch to the new wifi library
This removes the old wifi code entirely and instead links to the new
library at https://github.com/Pinoccio/library-gainspan-s2w. To use this
library after this commit, you should include that library in the
IDE libraries path as well.

Additionally, this new library is initialized to use SPI to talk to the
backpacks (which requires the SPI version of the firmware on the
gainspan module). To use UART for modules that still have the UART
firmware, see the comments in WifiBackpack::setup().

This commit disabled the wifi verbose mode for now, since the new wifi
library only has some compile-time verbosity switches.

Furthermore, the new wifi library does not yet set the L4 (TCP) retry
count for the HQ connection and does not wait for the association to
complete at startup.