Introduce a new Backpacks class
This class is responsible for keeping track of the various backpacks
connected to the scout. Right now, it only tracks their identifiers, but
this will be expanded to (parsed) EEPROM contents and Backpack object as
well later.

This removes the list of backpack identifiers from the PBBP class.
Instead, PBBP::enumerate() now gets a callback argument that is called
for every enumerated slave.