Use a union and struct for backpack unique ids
Instead of passing a raw byte array and relying on all code to pick out
the right bytes, this uses a struct with the right layout for passing
around unique ids. This allows code to just pick out the field they want
and the compiler will sort out what bytes they need.

The actual unique_id type is a union containing this struct as well as a
raw byte array, so the byte array can be accessed as well (without
breaking strict pointer aliasing rules, like casting the struct to
(uint8_t*) would probably do).