Restructure backpack EEPROM handling
Previously, there was some code in Backpacks::getEeprom() that fetched
the EEPROM contents. However, to do so it needed to know bits about the
EEPROM layout to know where the EEPROM size was stored. This code is now
moved inside the Pbbe class where it makes more sense (at the expense of
introducing a dependency from Pbbe onto PBBP.

Furthermore, instead of passing buffers and lengths everywhere, wrap
those in an Eeprom struct. Also slightly change the way the EEPROM
length is read and the EEPROM bytes are indexed.