Replace CRC implementations with template versions
This allows using the same code for 8-bit and 16-bit CRCs. This also
moves the function definitions to the header files and removes crc.cpp,
because they are now templates.

This removes the pinoccio_crc_generate_byte and
pinoccio_crc_generate_word functions in favour of a single
pinoccio_crc_generate function, requiring updating any external callers
(wifi test jig). The implementation of this function is also changed to
use pinoccio_crc_update instead of duplicating that code.