Don't put any static declarations in .h files
Static functions and variables are local to the compilation unit that
declares them, so putting them in .h files declares them in every
compilation unit that includes the .h file. If they are only used in a
single .cpp file, then they'll be optimized away from the other
compilation units, but might still trigger "declared but not defined" or
"declared but not used" compiler warnings.

References #63