Add a constructor to Pbbe::MajorMinor to fix warnings
The old implementation generated a warning:

warning: narrowing conversion of ‘(((int)rev) >> 4)’ from ‘int’ to ‘uint8_t {aka unsigned char}’ inside { } is ill-formed in C++11 [-Wnarrowing]

This is because even though rev is already uint8_t, it is autopromoted
to int when doing the shift, and the result is then truncated. This is
apparently ok in normal variable assignment, but will be invalid for
struct initializers since C++11, so gcc warns about it in non-C++11
mode. Also, the initializer format wasn't valid in non-C++11 mode in the first place,
which generated a second warning (#63).