Convert all sprintf calls to snprintf
This prevents writing outside of buffer boundaries, possibly truncating
strings (but C99 says snprintf should guarantee nul termination even
when the buffer is too small).

This commit just does a regex update of all calls so this is not very
smart about the length passed to snprintf (it duplicates a lot of strlen
callls that should be refactored later as well).