Don't let lead scouts broadcast their HQ reports
There is no need, if they send it to HQ directly then they shouldn't
also broadcast it. If we eventually support multiple lead scouts per
troup, we might need to revisit this part.

This commit seems to improve stability, but that's probably not because
it actually fixes things, but more that the reduced mesh traffic
prevents other problems from being exposed.