Let StringBuffer reserve memory block by block
Specifically, this adds a reserveBlock() method that does the same as
the regular reserve method (make sure the buffer size is at least the
value passed), but rounds the allocation size up to a multiple of the
block size (which is passed to the constructor).

Note that reserveBlock is only used by the StringBuffer methods and not
by the inherited String methods and operators. Specifically, this means
that using e.g. += will not do block reservation, but e.g. appendSprintf
will.