Add PrintToString class and printToString function
The class allows creating a Print object that prints to a string in
memory.

The function allows the same, but requires creating a new
function for every String printed to and requires the String to be
non-static global variables. This limits the usefulness of the function,
but until bitlash learns how to redirect its output to Print objects
instead of functions, this function should do.