Convert most buffers to use StringBuffer
This uses the StringBuffer object for most buffers in ScoutHandler.cpp
and Shell.cpp. The StringBuffer objects encapsulate all memory
management, automatically freeing up their memory when they go out of
scope.

When returning a StringBuffer object, the compiler automatically causes
the StringBuffer object itself to be allocated by the caller (return
value optimization) and the buffer is not copied (copy-elision) but just
passed to the caller and automatically freed in the caller.

When passing a buffer or string argument, some cases are modified to use
a const String&, but sometimes a const char* is still used, because
there isn't always a String object available in the caller (and creating
one currently always involved copying the entire string). This part
should be cleaned up a bit, but it's less important (since converting a
String or StringBuffer to char* is easy with the c_str() method).

This fixes #79.

Additionally, this commit moves the JSON escaping for commands ran by
field scouts from bitlash output function on the field scout to the
HQ-related functions on the lead scout, which make a lot more sense.

In addition to converting buffers, this also cleans up the buffer
handling in some places, which also fixes #86.