Let leadSignal accept a const String&
Every caller had a StringBuffer object that was converted to a char*, so
might a well pass the actual object.

Inside leadSignal, the String is now passed to the wifi client print
method instead of write. This means that the length stored in the
String object is used, saving a strlen call.