Pass a ConstBuf& instead of char* where possible
Previously, char* were passed, even when they were not guaranteed to be
nul-terminated (in particular, packets received over the mesh were not
guaranteed to be nul-terminated). Now, a ConstBuf object is passed,
which includes the length, preventing out-of-bounds reads.

This removes most of the c_str() calls from ScoutHandler. The ones
remaining are for interfacing with external things, like bitlash, js0n
and sprintf, which we cannot change to accept anything else.