Fix escaping in appendJsonString
Previously, appendJsonString kept a char *out, pointing to the next
character to write that was valid throughout the entire function.
However, when characters needed to be escaped, usually the reserved
buffer would be too small and a second reallocation was needed. This
could change the address of the buffer, making the out pointer invalid.

To fix this, reset the out pointer at the start of each try and keep an
explicit count of bytes written so we know where to point it.