Remove bitlashFilter function
This function was used to redirect bitlash output to in the normal case
(e.g. when not running HQ commands). However, this function simply
prints to Serial, so it makes more sense to just reset the output
handler and let bitlash print to Serial itself.

This function was probably a remnant from when things were structured
differently.