Add 'power.hibernate' function to put the scout to sleep
This function disables the mesh networking and as much hardware as
possible and puts the chip into "deep sleep".

It uses timer2, controlled by the 32khz clock to schedule the wakeup. If
any other interrupt occurs in the meanwhile, the hibernate could be cut
short (but most other interrupts are not able to actually wake up the
scout from deep sleep).

After the hibernate is complete, an optional callback bitlash function
is called.

Going into hibernation does not yet disable any backpacks and leaves the
backpack vcc on. Serial communiation might be cut off halfway through a
byte, this should be improved later.