Add constants for pin modes
There were already some constants, but these were a mix of the Arduino
pin mode constants and some extra defines). This unifies this to a
separate enum with PINMODE_* constants, which are used everywhere
Pinoccio pinmodes are used.

In setMode(), the Pinoccio and Arduino modes are still interchanged,
this will be fixed next.