Improve checkArgs to support a maximum
Previously, checkArgs could either check for an exact number of
arguments or at least some number of arguments, triggered by a messy
boolean argument.

Now, there are two overloaded versions of the function, one that checks
for an exact number of arguments and one that checks for both a minimum
and maximum number of arguments.

This allows stricter checking - some functions would previously not
raise an error when too many arguments were passed.