Don't mix pinmode constants in PinoccioScout::setMode
Previously, the Pinoccio pin modes were sometimes used as Arduino
pinmodes. This worked, because INPUT/OUTPUT/INPUT_PULLUP overlap, but
it's not very elegant. Now, a switch is used to map all the supported
PINMODE_* constants into the proper pinMode argument instead.

Also, the pin value is now also determined in a switch, to prevent
calling pinRead while it's not actually needed. As an extra advantage,
this calls pinRead after changing the pinmode, so when changing to
OUTPUT, the value read should better reflect the actual output value.