Make isPinReserved work for all pins
Currently, the function is only called for the Dx and Ax pins. It used
to only look at backpacks for deciding about reserved pins, but now it
also knows about the pins that are connected to something on the
Pinoccio Scout board itself (fuel gauge, serial port, led, etc).

These pins are hardcoded for now, this should be generalized later on.