Add key.signed/key.unsigned commands
key.signed() is a new command to create a key string from a signed integer value.  key.unsigned() is the complementary function for unsigned values, but is simply an alias for key(), which already handles string AND unsigned int arguments.  key.number() (unchanged) will convert either signed or unsigned integer strings to numeric values for comparisons.