Allow multiple args to hq.Report.  Fix key leak
This allows multiple arguments to be passed to hq.Report().  The functionality was all there, but the checkArgs() call at the top of hqReport() was checking for exactly 2 arguments.

Also noticed a bug in arg2array.  It was allocating a "sticky" key for each string argument but not freeing them.  I changed the "at" flag from 0 to 1 to allow keyLoop to automatically free the keys on each Bitlash loop.  hq.Report still reports the correct data but now it won't fill the key table.