Write the sleep end time to the compare register sooner
Instead of storing the sleep end time in a variable for a while and
writing it to the symbol counter compare register just before sleeping,
we now write it to the register right away. This makes it easier to
detect if the sleep time already passed while waiting for unfinished
business.
Since the hardware now keeps a flag for this, we no longer
have to check for overflow and we can again juse the full uin32_t range,
extending the maximum sleep time to 19 hours.