Let pin.save actually save again
The startup function that stores the pinmode and value was still
generated, but in commit b33ab26 (trying to fix the bitlash re-entrant
issue) the actual doCommand call that saves it was (accidentally?)
removed. pin.save would still set the pin mode, but it would not be
saved across reboots.