Fix lockup in SleepHandler::doSleep
After enabling the ADC after sleeping, a loop waited for VDD to be back
at the right level before continuing. However, the condition on the loop
was inverted, meaning the loop would wait _while_ the level was correct.

If the VDD level was ok before this loop was reached, this would lock up
the scout. In practice, it seems this didn't happen until wakeup
interrupts were added (which add some delay after sleep).