Make pin.save work with pin numbers too
Before, when passed a pin number instead of a name, it would set the pin
mode but then fail to create the corresponding startup function since it
had no name available. Now, it looks up the name for the pin number
instead of using the string argument directly. Since this name is a
flash string, the previous printf approach doesn't work anymore, so we
use a StringBuffer instead (which allows easily concatenating flash
strings).