Don't call updatePinState for unused pins periodically
Previously, updatePinSate was called when the pin mode was changed, and
periodically from the state change handler, even for unused pins. This
meant updatePinState had to skip the event handlers and updating the
state for unused pins, _except_ when the mode changed (in case the mode
changed from used to unused).

This latter is needlessly complex and can be prevented by simply not
calling updatePinState for unused pins periodically. Now updatePinState
no longer ignores unused pins - if it's called for an unused pin, then
it's for a pin change so it should do its work anyway.

As an added bonus, this prevents doing a pinRead for unused pins, whose
result isn't actually used later anyway.