Add pin.number bitlash function
This function returns the pin number for a given pin name, allowing
things like using a variable to store pin numbers based on a name. If a
number is passed to pin.number, it is checked for being a valid pin
number. If an invalid name or numbers is passed, -1 is returned.