Stop supporting a full command as callback to power.sleep
Before, you could pass a full bitlash command as the callback for
power.sleep. In preparation of passing arguments to the callback, this
support is removed and only a plain function name can be passed now.

This commit only adds a single check and renames variables, it does not
otherwise modify any behaviour.