Resume interrupted sleeps when the callback returns non-zero
Previously when a sleep was interrupted, the callback was ran and the
sleep ended. This is still the default, but if the callback returns a
non-zero value, the sleeping continues (until interrupted again or
finished, the callback is ran again in either case).

Note that before doing the actual sleeping, the loop functions run at
least once, possibly longer if they create unfinished business that the
sleep code waits for.