Add module support
A Module is a class the defines a bunch of functionality that can be
dynamically enabled at runtime. The code is already present in Flash,
but the relevant setup steps (in particular, big memory allocations)
have not been performed and the associated bitlash function are not
registered yet until the module is enabled.

Each module should define a single instance of itself (and make its
constructor private to prevent creating further instances), which will
automatically register itself into the linked list of available modules
at startup (through the Module class constructor). This means there is
no need to keep a central list of available modules, and any modules
that live in external Arduino libraries are automatically included too.

The new module-related classes are put into the new "pinoccio"
namespace, to prevent potential naming conflicts with other code. The
rest of the code should be converted to use this namespace later, but
using the namespace for this new code right away probably makes things
easier.

Disabling a module is not currently supported, mainly because bitlash
has no way to un-register functions.