Stop using NCM for TCP connections to HQ
NCM seems to be a bit dodgy. Also, there is a bug in the Gainspan
firmware where you can never shorten the hostname that NCM uses (it
never writes a trailing zero, but writes into a buffer filled with
zeroes on factory reset), which we now circumvent.

To make sure that we properly reconnect if the TCP connection breaks for
whatever reason, we just check if we're still connected in every loop.
To prevent a race condition between the TCP connection and the timesync
(which happens in the onAssociate event handler), a flag is kept to
indicate if time sync ever happened succesfully.