Let WifiModule::bp() return a pointer
Previously it returned a reference, but returning a pointer allows it to
return NULL next.

Additionally, this changes all the calling code to check the bp()
pointer instead of the module loaded() status before using the backpack
methods. Currently, there is no way for bp() to return NULL when the
backpack is loaded, but this will change in the future.