Let the backpack bus stuff handle backpack VCC toggling
When the backpack vcc is turned off, the list of backpacks is cleared.
When it is turned back on, the backpack bus is re-initialized.

Before this change, toggling the power would cause the backpack bus to
stop working, since the slaves forget their assigned addresses when
power cycled, but the 256rfr2 expected them to still respond to their
previously assigned address. By re-enumerating the bus, this is solved.