Finish up Shell::eval
It was previously merged in a half-finished state. Now, there is only a
single eval method with a few overloads (evalArgs was removed). See
Shell.h for examples on how to use it.

The Concatenator.h file is also cleaned up and documented, slightly
changing it as well (it now concats into all String objects (including
StringBuffer) instead of just into StringBuffer).