Add a "sniffer" module
This module allows capturing packets from the mesh. To use it, enable the
module and call sniffer.start:

    > module.enable("sniffer")
    > sniffer.start

This dumps raw packets to the Serial port (but completely stops all
other processing on the Scout!).

By running `sniffer.start(1)`, a binary output mode is selected which
works in concert with a small script:

https://github.com/Pinoccio/tool-serial-pcap