Fix off-by-one in command ack handling
This caused the ack command to overflow the command buffer, as well as
caused the ack command to be read from outside the buffer. This was
visible in the form of an "unexpected char" error message on a lead
scout after it connected to HQ, due to the "hq.online(0)" command it
broadcasted.