Add a queue to the sniffer module
A big packet takes a (relatively) long time to transmit over serial.
Since the radio is about twice as fast (raw bitrate) as serial, a big
packet could keep the CPU busy so long that it didn't get around to
reading new packets as they came in.

This commit adds a queue so packets are copied to the queue quickly when
they come in and then sent byte-by-byte to serial, checking for new
packets after every byte.

This still means that packets will be dropped if a packets are sustained
at a high rate. However, the queue takes up all the free memory, so
bursts as long as possible can still be captured.