ota: Resend the right data on tx retry
Previously, when a data command didn't get ack'd, the code would "retry"
with the next chunk of data, making the retry essentially useless. This
commit waits with updating the counters until the packet is ack'd, so
the right data is resent on tx failure.

Note that in practice, it seems that a missing ack and retry is often
followed by a crc error, so it might be that the data packet is received
ok and it's just the ack that gets missing. This needs some further
testing to find the best strategy.