Merge pin mode and state arrays
This uses a single array of structs. By using bitfields, we shrink the
value to 10 bits (needed for analog measurements) and shrink the mode to
4 bits, allowing them both to fit into 2 bytes instead of 3.

This does mean that negative values are no longer possible, but these
were only set for the negative modes, in which case the value wasn't
used in the first place.