Duplicate KEYS_BUNDLE before loading it
keyLoad modifies the string passed to it during JSON parsing. Passing a
string literal is therefore not a good idea, since those shouldn't be
modified. On AVR, there is no read-only memory to prevent modification,
so it didn't show up as a problem before.

However, when you try to use the string "\"]" in the code, this will
share memory with the last two bytes of KEYS_BUNDLE. This means that if
KEYS_BUNDLE is modified, that two-byte string will also be modified,
leading to unexpected results.

To prevent this, just strdup KEYS_BUNDLE before loading it.