lora: Increase the reception timeout
Since the SX1272 library only makes the hardware listen during these
timeouts and the entire message must be received and processed in that
time, having a tiny timeout doesn't work at all, or only sporadically.

This approach means that the loop might block up to 1000ms, which will
break other stuff, so the SX1272 library needs to be changed (or
replaced) to allow for more asynchronous behaviour. For now, this longer
timeout makes at least some things work, though.