Include lwm.h earlier in Bootstrap.ino
Recent Arduino IDE versions changed the way that dependencies are
detected. Instead of looking just at the .ino file with regexes, the IDE
now runs the preprocessor and collects errors. Since only top-level
includes from a library cause it to be included, but some .h files in
this library include only lower level include files (and those would be
processed before the top-level include in Bootstrap.ino), this would
result in an include error. By adding a top-level lwm.h include to a few
places, this is fixed.

This fixes https://github.com/Pinoccio/firmware-pinoccio/issues/13