Fix compile error due to bit shift in constant
The 1 defaults to int, which is signed, so shifting it by 31 bits is too
much. This probably was a warning before, and then it would probably
work fine because the result is cast to an unsigned int, but with recent
gcc/arduino versions this produces a compilation error.

By casting to unsigned first, this error is avoided.